requirements.txt
#python=3.10,windows,conda env
pandas
numpy
python-docx
bayoo-docx
python-dotenv